requirements.txt
opencv-python==4.10.0.84
pillow==10.4.0
streamlit==1.38.0
streamlit-image-coordinates==0.1.9